requirements.txt
comtypes
python-pptx-interface 
edge-tts
mutagen
opencv-python 
moviepy
pandas
